pyproject.toml
[project]
name = "slack-agent"
version = "0.1.0"
description = "Slack Agent built with Google Agent Development Kit"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk==0.4.0",
    "slack-bolt>=1.23.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.11.16",
    "litellm>=1.67.0",
]

[dependency-groups]
dev = [
    "pyright>=1.1.399",
    "ruff>=0.11.4",
]

[build-system]
requires = ["setuptools", "wheel", "pip"]
build-backend = "setuptools.build_meta"
